clean the slate
