--- packages/commonwealth/client/styles/pages/notification_settings.scss
@@ -1,104 +1,0 @@
-@import 'client/styles/shared';
-
-.NotificationSettingsPage {
-  padding: 24px;
-
-  .EmailIntervalConfiguration {
-    background: $neutral-100;
-    padding: 18px 24px 8px;
-    border-radius: 5px;
-    margin-top: 8px;
-    margin-bottom: 40px;
-    display: flex;
-    h4 {
-      margin: 0;
-    }
-    p {
-      margin: 3px 0 15px;
-    }
-    .cui-radio {
-      margin-bottom: 15px;
-    }
-    .email-interval-configuration-left,
-    .email-interval-configuration-right {
-      padding-right: 30px;
-    }
-    @include smallInclusive {
-      .email-interval-configuration-left {
-        margin-bottom: 20px;
-      }
-    }
-  }
-  .NotificationsTable {
-    margin-top: 30px;
-    margin-bottom: 40px;
-    th {
-      font-weight: 500;
-      padding: 6px 0 12px;
-      border-bottom-color: $neutral-100;
-    }
-    td {
-      border: none;
-      padding: 6px 0;
-      height: 43px;
-      &.subscription-label {
-        padding-right: 30px;
-        .cui-tag {
-          margin-left: 8px;
-        }
-        .ChainEventDetails {
-          color: $neutral-500;
-          font-size: 0.8em;
-        }
-      }
-      .item-metadata {
-        font-size: 90%;
-        color: $neutral-500;
-        margin-left: 6px;
-      }
-    }
-    tr:nth-child(2) td {
-      padding-top: 12px;
-    }
-    tr.on-chain-events-header + tr {
-      td {
-        padding-top: 12px;
-      }
-    }
-    tr.on-chain-events-header th {
-      padding-top: 36px;
-    }
-    th.subscription-setting,
-    td.subscription-setting {
-      text-align: right;
-      button {
-        width: 170px;
-        text-align: left;
-      }
-      button.sometimes {
-        color: $neutral-500;
-      }
-    }
-  }
-}
-
-// popovers
-.CommunityNotificationSelectList,
-.BatchedNotificationSelectList,
-.ChainNotificationSelectList,
-.EventSubscriptionTypeSelectList {
-  .cui-popover-content {
-    padding: 0;
-    min-width: 125px;
-    overflow: scroll;
-  }
-  .cui-list-item.cui-active {
-    background: none;
-  }
-  .cui-list-item.cui-selected {
-    background: $primary-300;
-  }
-  .cui-list-item:hover {
-    background: $primary-300 !important;
-  }
-}
