+++ packages/commonwealth/client/styles/pages/notification_settings/index.scss
@@ -1,0 +1,6 @@
+@import 'shared';
+
+.NotificationSettingsPage {
+  flex: 1;
+  padding: 24px;
+}
